# server/go/model_tasks_for_req.go
/*
 * Todo Manager API
 *
 * This is todo manager backend for coding tests.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type TasksForReq struct {

	Title string `json:"title,omitempty"`

	Checked bool `json:"checked,omitempty"`

	Description string `json:"description,omitempty"`
}

// AssertTasksForReqRequired checks if the required fields are not zero-ed
func AssertTasksForReqRequired(obj TasksForReq) error {
	return nil
}

// AssertRecurseTasksForReqRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of TasksForReq (e.g. [][]TasksForReq), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseTasksForReqRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aTasksForReq, ok := obj.(TasksForReq)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertTasksForReqRequired(aTasksForReq)
	})
}