# server/go/model_tasks_for_res.go
/*
 * Todo Manager API
 *
 * This is todo manager backend for coding tests.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type TasksForRes struct {

	Id string `json:"id,omitempty"`

	Title string `json:"title,omitempty"`

	Checked bool `json:"checked,omitempty"`

	Description string `json:"description,omitempty"`

	SettelementMillis int32 `json:"settelementMillis,omitempty"`
}

// AssertTasksForResRequired checks if the required fields are not zero-ed
func AssertTasksForResRequired(obj TasksForRes) error {
	return nil
}

// AssertRecurseTasksForResRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of TasksForRes (e.g. [][]TasksForRes), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseTasksForResRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aTasksForRes, ok := obj.(TasksForRes)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertTasksForResRequired(aTasksForRes)
	})
}