# server/go/model_users_for_req.go
/*
 * Todo Manager API
 *
 * This is todo manager backend for coding tests.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type UsersForReq struct {

	Name string `json:"name,omitempty"`

	EmailAddress string `json:"emailAddress,omitempty"`

	PhoneNumber string `json:"phoneNumber,omitempty"`
}

// AssertUsersForReqRequired checks if the required fields are not zero-ed
func AssertUsersForReqRequired(obj UsersForReq) error {
	return nil
}

// AssertRecurseUsersForReqRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of UsersForReq (e.g. [][]UsersForReq), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseUsersForReqRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aUsersForReq, ok := obj.(UsersForReq)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertUsersForReqRequired(aUsersForReq)
	})
}